Fix dev problem (app dying on hot reload); changes to Cloze reveal
- Add STARTUP_TIMEOUT to custom/constants.js

On hot reload, various components were already loaded, but the connection to the collections was broken, so findOne() calls were returning undefined and subsequent operations were failing. Also, Session "group_id" was being unset, so App tracker was serving props with no `page` property. This meant that the Splash screen was shown again, and StartUp was called but then timed out.

- Save view sent to hideSplash in App state, to replace "Splash"
- Check if Menu item data is undefined before trying to create item
- Remove unnecessary subscriptions property from StartUp
- ClozeCore
  - Add revealDelay and revealTimeout properties
  - Edit how reveal is added to Group page.data
  - Add revealed and show to Group page.data
  - Set timeout to reveal again after input pause without typing
  - Add this.submit and this.revealAnswer to props for components
  - Don't treatInput for input changes fromNewPhrase
- Cloze inputs: modify treatment of this.props